requirements.txt
black==19.3b0
Django>=2.2,<2.3
django-environ>=0.4.5,<0.5
django-filter>=2.2.0,<2.3
djangorestframework>=3.10.3,<3.11
drf-yasg>=1.17.0,<1.18.0
gunicorn>=19.9.0,<19.10.0
json-logging-py==0.2
mysqlclient>=1.4.4,<1.5
pyflakes==2.1.1
PyYAML>=5.1.2,<5.2
six>=1.12.0,<1.13
uritemplate>=3.0.0,<3.1
whitenoise>=4.1.4,<4.2.0
git+https://github.com/HumanExposure/factotum.git@dev#egg=dashboard

